+++ example/lib/03.change_language_theme/constants.dart
@@ -1,0 +1,36 @@
+import 'package:highlight/languages/dart.dart';
+import 'package:highlight/languages/go.dart';
+import 'package:highlight/languages/java.dart';
+import 'package:highlight/languages/php.dart';
+import 'package:highlight/languages/python.dart';
+import 'package:highlight/languages/scala.dart';
+import 'package:highlight/languages/yaml.dart';
+
+final builtinLanguages = {
+  'dart': dart,
+  'go': go,
+  'java': java,
+  'php': php,
+  'python': python,
+  'scala': scala,
+  'yaml': yaml,
+};
+
+const languageList = <String>[
+  'dart',
+  'go',
+  'java',
+  'php',
+  'python',
+  'scala',
+  'yaml',
+];
+
+const themeList = <String>[
+  'a11y-dark',
+  'an-old-hope',
+  'atom-one-dark',
+  'monokai-sublime',
+  'vs',
+  'vs2015',
+];