+++ lib/src/code_field/code_controller/code_controller.dart
@@ -161,6 +161,8 @@ class CodeController extends TextEditingController {
   @visibleForTesting
   TextSpan? lastTextSpan;
 
+  bool _disposed = false;
+
   late final actions = <Type, Action<Intent>>{
     CommentUncommentIntent: CommentUncommentAction(controller: this),
     CopySelectionTextIntent: CopyAction(controller: this),
@@ -191,6 +193,10 @@ class CodeController extends TextEditingController {
     if (_code.text != codeSentToAnalysis.text) {
       // If the code has been changed before we got analysis result, discard it.
       // This happens on request race condition.
+      return;
+    }
+
+    if (_disposed) {
       return;
     }
 
@@ -778,6 +784,7 @@ class CodeController extends TextEditingController {
 
   @override
   void dispose() {
+    _disposed = true;
     _debounce?.cancel();
     historyController.dispose();
     _suggestionHelper.dispose(); // Clean up the isolate when the controller is disposed