+++ lib/src/code_field/code_field.dart
@@ -95,6 +95,7 @@ class CodeField extends StatefulWidget {
   const CodeField({
     super.key,
     required this.controller,
+    this.undoController,
     this.minLines,
     this.maxLines,
     this.expands = false,
@@ -138,6 +139,10 @@ class CodeField extends StatefulWidget {
   /// A CodeController instance to apply
   /// language highlight, themeing and modifiers.
   final CodeController controller;
+
+  /// An UndoHistoryController instance
+  /// to control TextField history.
+  final UndoHistoryController? undoController;
 
   @Deprecated('Use gutterStyle instead')
   final GutterStyle lineNumberStyle;
@@ -224,6 +229,9 @@ class _CodeFieldState extends State<CodeField> {
     _codeScroll?.addListener(_updatePopupOffsetOnScroll);
 
     WidgetsBinding.instance.addPostFrameCallback((_) {
+      if (!mounted) {
+        return;
+      }
       final double width = _codeFieldKey.currentContext!.size!.width;
       final double height = _codeFieldKey.currentContext!.size!.height;
       windowSize = Size(width, height);
@@ -270,6 +278,9 @@ class _CodeFieldState extends State<CodeField> {
   void rebuild() {
     setState(() {
       WidgetsBinding.instance.addPostFrameCallback((_) {
+        if (!mounted) {
+          return;
+        }
         // For some reason _codeFieldKey.currentContext is null in tests
         // so check first.
         final context = _codeFieldKey.currentContext;
@@ -362,6 +373,7 @@ class _CodeFieldState extends State<CodeField> {
     final defaultTextStyle = TextStyle(
       color: styles?[rootKey]?.color ?? DefaultStyles.textColor,
       fontSize: themeData.textTheme.titleMedium?.fontSize,
+      height: themeData.textTheme.titleMedium?.height,
     );
 
     textStyle = defaultTextStyle.merge(widget.textStyle);
@@ -378,6 +390,7 @@ class _CodeFieldState extends State<CodeField> {
       scrollPadding: widget.padding,
       style: adjustedTextStyle,
       controller: widget.controller,
+      undoController: widget.undoController,
       minLines: widget.minLines,
       maxLines: widget.maxLines,
       expands: widget.expands,