+++ lib/src/single_line_comments/parser/single_line_comments.dart
@@ -2,10 +2,13 @@ import 'package:highlight/highlight_core.dart';
 import 'package:highlight/languages/dart.dart';
 import 'package:highlight/languages/go.dart';
 import 'package:highlight/languages/java.dart';
+import 'package:highlight/languages/javascript.dart';
 import 'package:highlight/languages/php.dart';
 import 'package:highlight/languages/python.dart';
 import 'package:highlight/languages/scala.dart';
+import 'package:highlight/languages/typescript.dart';
 import 'package:highlight/languages/vhdl.dart';
+import 'package:highlight/languages/yaml.dart';
 
 class SingleLineComments {
   const SingleLineComments._();
@@ -18,6 +21,9 @@ class SingleLineComments {
     python: [_hash],
     scala: [_slashes],
     vhdl: [_hyphenMinuses],
+    typescript: [_slashes],
+    javascript: [_slashes],
+    yaml: [_hash],
   };
 
   static const _slashes = '//';